refactor: gpflow utils
