+++ queens/models/logpdf_gp.py
@@ -28,7 +28,7 @@ from scipy import stats
 
 from queens.models.model import Model
 from queens.utils import jax_minimize_wrapper
-from queens.utils.gpflow import init_scaler
+from queens.utils.gpflow_transformations import init_scaler
 from queens.utils.numpy_linalg import safe_cholesky
 
 _logger = logging.getLogger(__name__)
