+++ queens/models/surrogates/variational_gaussian_process.py
@@ -37,8 +37,9 @@ import tensorflow_probability as tfp
 
 from queens.models.surrogates.surrogate import Surrogate
 from queens.utils.configure_tensorflow import configure_keras, configure_tensorflow
-from queens.utils.gpflow import extract_block_diag, init_scaler, set_transform_function
+from queens.utils.gpflow_transformations import init_scaler, set_transform_function
 from queens.utils.logger_settings import log_init_args
+from queens.utils.numpy_array import extract_block_diag
 
 _logger = logging.getLogger(__name__)
